requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# friends-word-rnn.py: 13,14,15,16,17,18
Keras == 2.0.9

# friends-word-rnn.py: 9
numpy == 1.13.1

# friends-word-rnn.py: 8
tensorflow_gpu == 1.3.0
